upgraded message broker logic
