+++ src/extract/extract.py
@@ -7,7 +7,6 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../'
 import logging
 import json
 import pika
-import ssl
 import time
 import fitz
 from pathlib import Path
@@ -21,6 +20,8 @@ from docx import Document
 from io import BytesIO
 from llama_index import download_loader
 from shared.vectorflow_request import VectorflowRequest
+from services.rabbitmq.rabbit_service import create_connection_params
+from pika.exceptions import AMQPConnectionError
 
 logging.basicConfig(filename='./extract-log.txt', level=logging.INFO)
 logging.basicConfig(filename='./extract-error-log.txt', level=logging.ERROR)
@@ -172,30 +173,16 @@ def callback(ch, method, properties, body):
 
     ch.basic_ack(delivery_tag=method.delivery_tag)
 
-def start_connection():
+def start_connection(max_retries=5, retry_delay=5):
     global publish_channel
     global connection
-    
-    while True:
+
+    for attempt in range(max_retries):
         try:
-            credentials = pika.PlainCredentials(os.getenv('RABBITMQ_USERNAME'), os.getenv('RABBITMQ_PASSWORD'))
-
-            connection_params = pika.ConnectionParameters(
-                host=os.getenv('RABBITMQ_HOST'),
-                credentials=credentials,
-                port=os.getenv('RABBITMQ_PORT'),
-                heartbeat=600,
-                ssl_options=pika.SSLOptions(ssl.SSLContext(ssl.PROTOCOL_TLSv1_2)),
-                virtual_host="/"
-            ) if os.getenv('RABBITMQ_PORT') == "5671" else pika.ConnectionParameters(
-                host=os.getenv('RABBITMQ_HOST'),
-                credentials=credentials,
-                heartbeat=600,
-            )
-
+            connection_params = create_connection_params()
             connection = pika.BlockingConnection(connection_params)
             consume_channel = connection.channel()
-            publish_channel = connection.channel() 
+            publish_channel = connection.channel()
 
             consume_queue_name = os.getenv('EXTRACTION_QUEUE')
             publish_queue_name = os.getenv('EMBEDDING_QUEUE')
@@ -207,10 +194,27 @@ def start_connection():
 
             logging.info('Waiting for messages.')
             consume_channel.start_consuming()
-            
+            return  # If successful, exit the function
+
+        except AMQPConnectionError as e:
+            logging.error('AMQP Connection Error: %s', e)
         except Exception as e:
-            logging.error('ERROR connecting to RabbitMQ, retrying now. See exception: %s', e)
-            time.sleep(10) # Wait before retrying
+            logging.error('Unexpected error: %s', e)
+        finally:
+            if connection and not connection.is_closed:
+                connection.close()
+
+        logging.info('Retrying to connect in %s seconds (Attempt %s/%s)', retry_delay, attempt + 1, max_retries)
+        time.sleep(retry_delay)
+
+    raise Exception('Failed to connect after {} attempts'.format(max_retries))
 
 if __name__ == "__main__":
-    start_connection()
+    while True:
+        try:
+            start_connection()
+
+        except Exception as e:
+            logging.error('Error in start_connection: %s', e)
+            logging.info('Restarting start_connection after encountering an error.')
+            time.sleep(10)
