+++ src/worker/vdb_upload_worker.py
@@ -4,15 +4,11 @@ import os
 # this is needed to import classes from the API. it will be removed when the worker is refactored
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../')))
 
-import ssl
 import time
 import pika
 import json
-import openai
 import pinecone
-import requests 
 import logging
-import uuid
 import weaviate
 import redis
 import lancedb
@@ -32,6 +28,8 @@ from shared.embeddings_type import EmbeddingsType
 from shared.vector_db_type import VectorDBType
 from shared.utils import generate_uuid_from_tuple
 from urllib.parse import quote_plus
+from services.rabbitmq.rabbit_service import create_connection_params
+from pika.exceptions import AMQPConnectionError
 
 logging.basicConfig(filename='./vdb-upload-log.txt', level=logging.INFO)
 logging.basicConfig(filename='./vdb-upload-errors.txt', level=logging.ERROR)
@@ -412,24 +410,10 @@ def callback(ch, method, properties, body):
 
     ch.basic_ack(delivery_tag=method.delivery_tag)
 
-def start_connection():
-    while True:
+def start_connection(max_retries=5, retry_delay=5):
+    for attempt in range(max_retries):
         try:
-            credentials = pika.PlainCredentials(os.getenv('RABBITMQ_USERNAME'), os.getenv('RABBITMQ_PASSWORD'))
-
-            connection_params = pika.ConnectionParameters(
-                host=os.getenv('RABBITMQ_HOST'),
-                credentials=credentials,
-                port=os.getenv('RABBITMQ_PORT'),
-                heartbeat=600,
-                ssl_options=pika.SSLOptions(ssl.SSLContext(ssl.PROTOCOL_TLSv1_2)),
-                virtual_host="/"
-            ) if os.getenv('RABBITMQ_PORT') == "5671" else pika.ConnectionParameters(
-                host=os.getenv('RABBITMQ_HOST'),
-                credentials=credentials,
-                heartbeat=600,
-            )
-
+            connection_params = create_connection_params()
             connection = pika.BlockingConnection(connection_params)
             channel = connection.channel()
 
@@ -440,11 +424,24 @@ def start_connection():
 
             logging.info('Waiting for messages.')
             channel.start_consuming()
+            return  # If successful, exit the function
             
+        except AMQPConnectionError as e:
+            logging.error('AMQP Connection Error: %s', e)
         except Exception as e:
-            logging.error('ERROR connecting to RabbitMQ, retrying now. See exception: %s', e)
-            time.sleep(config.PIKA_RETRY_INTERVAL) # Wait before retrying
+            logging.error('Unexpected error: %s', e)
+        finally:
+            if connection and not connection.is_closed:
+                connection.close()
+
+        logging.info('Retrying to connect in %s seconds (Attempt %s/%s)', retry_delay, attempt + 1, max_retries)
+        time.sleep(retry_delay)
 
 if __name__ == "__main__":
-    start_connection()
- 
+    while True:
+        try:
+            start_connection()
+        except Exception as e:
+            logging.error('Error in start_connection: %s', e)
+            logging.info('Restarting start_connection after encountering an error.')
+            time.sleep(config.PIKA_RETRY_INTERVAL)
